requirements.txt
# Flask Framework
Flask==3.0.0
Werkzeug==3.0.1

# Database
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23

# Authentication
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1

# Password Hashing
bcrypt==4.1.1

# WSGI Server (for production)
gunicorn==21.2.0

# Environment Variables
python-dotenv==1.0.0

# Utilities
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3

# Additional dependencies
blinker==1.7.0
typing-extensions==4.9.0
